Tweaked shooting and bullets

+++ src/com/nak/starfire/utilities/Time.java
@@ -3,17 +3,17 @@ package com.nak.starfire.utilities;
 public class Time {
 
 	public static final long SECOND = 1000000000L;
-	public static double delta;
+	public static long delta;
 
 	public static long getTime() {
 		return System.nanoTime();
 	}
 
-	public static double getDelta() {
+	public static long getDelta() {
 		return delta;
 	}
 
-	public static void setDelta(double delta) {
+	public static void setDelta(long delta) {
 		Time.delta = delta;
 	}
